Move CidrRange into its own files. Move truncateIpAddressAndLength into CidrRange as a static file. Add getIpv4AnyAddress and getIpv6AnyAddress into Network::Utility. Move decl of parseInternetAddress from address.h to address_impl.h. Mark various single arg constructors as "explicit" per style guide.
