+++ test/CMakeLists.txt
@@ -72,6 +72,7 @@ add_executable(envoy-test
   common/mongo/proxy_test.cc
   common/mongo/utility_test.cc
   common/network/address_impl_test.cc
+  common/network/cidr_range_test.cc
   common/network/connection_impl_test.cc
   common/network/dns_impl_test.cc
   common/network/filter_manager_impl_test.cc
